+++ base_extended_security/controllers/__init__.py
@@ -1,4 +1,8 @@
 # © 2019 Numigi (tm) and all its contributors (https://bit.ly/numigiens)
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 
-from . import main, web_export
+from . import (
+    crud,
+    search,
+    web_export,
+)
