+++ base_extended_security/tests/__init__.py
@@ -1,4 +1,2 @@
 # © 2019 Numigi (tm) and all its contributors (https://bit.ly/numigiens)
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
-
-from . import test_controllers
